more final stuff
Game code includes array of ellipses that fall at different speeds, movable rectangle. Still needs interaction between rectangle and ellipses. When elipse touches the rectangle it should turn black (illusions of disappearence, I've got them!)

The ellipses are not written as a class... hopefully thats not a huge deal that hinders the interaction.